GeV5_refactor/requirements.txt
flask
requests
numpy
matplotlib
pyserial
websocket-client
waitress
Werkzeug
RPi.GPIO
fastapi
uvicorn
reportlab
